Create ParseURL.kt
You want to hack a website and get all available parameters. Print them "key : value". If a parameter doesn't have value, print "not found".

If you find the password (parameter pass), you should print its value after all parameters once again, but with a key password. If a URL does not contain a parameter pass, do not print anything after parameters. But if the pass parameter is present, its value cannot be empty

Note: the order of parameters should be like as in the URL.
Advice: learn the structure of URL.